+++ src/main/java/com/imsweb/algorithms/yostacspoverty/YostAcsPovertyInputDto.java
@@ -1,47 +1,30 @@
 package com.imsweb.algorithms.yostacspoverty;
 
-public class YostAcsPovertyInputDto {
+import com.imsweb.algorithms.StateCountyTractInputDto;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.math.NumberUtils;
 
-    private String _addressAtDxState;
+import static com.imsweb.algorithms.yostacspoverty.YostAcsPovertyUtils.YOST_ACS_CENSUS_TRACT_PIVOT_YEAR;
 
-    private String _countyAtDxAnalysis;
+public class YostAcsPovertyInputDto extends StateCountyTractInputDto {
 
-    private String _censusTract2010;
-
-    private String _censusTract2020;
+    private static final String MISSING_YEAR = "";
 
     private String _dateOfDiagnosis;
+    private String _yearOfDiagnosis;
 
-    public String getAddressAtDxState() {
-        return _addressAtDxState;
+    @Override
+    public void applyRecodes() {
+        super.applyRecodes();
+        _dateOfDiagnosis = _dateOfDiagnosis == null ? "" : _dateOfDiagnosis.trim();
+        _yearOfDiagnosis = _yearOfDiagnosis == null ? "" : _yearOfDiagnosis.trim();
     }
 
-    public void setAddressAtDxState(String addressAtDxState) {
-        _addressAtDxState = addressAtDxState;
-    }
-
-    public String getCountyAtDxAnalysis() {
-        return _countyAtDxAnalysis;
-    }
-
-    public void setCountyAtDxAnalysis(String countyAtDxAnalysis) {
-        _countyAtDxAnalysis = countyAtDxAnalysis;
-    }
-
-    public String getCensusTract2010() {
-        return _censusTract2010;
-    }
-
-    public void setCensusTract2010(String censusTract2010) {
-        _censusTract2010 = censusTract2010;
-    }
-
-    public String getCensusTract2020() {
-        return _censusTract2020;
-    }
-
-    public void setCensusTract2020(String censusTract2020) {
-        _censusTract2020 = censusTract2020;
+    public int computeYearOfDiagnosis() {
+        if (isValidYear(_yearOfDiagnosis)) {
+            return Integer.parseInt(_yearOfDiagnosis);
+        }
+        return -1;
     }
 
     public String getDateOfDiagnosis() {
@@ -50,5 +33,45 @@ public class YostAcsPovertyInputDto {
 
     public void setDateOfDiagnosis(String dateOfDiagnosis) {
         _dateOfDiagnosis = dateOfDiagnosis;
+        _yearOfDiagnosis = dateOfDiagnosis == null ? null : StringUtils.rightPad(dateOfDiagnosis, 4).substring(0, 4).trim();
     }
+
+    public boolean hasInvalidStateCountyCensusTractOrYear() {
+        if (isInvalidState(getAddressAtDxState()) || isInvalidCounty(getCountyAtDxAnalysis()) || isInvalidYear(_yearOfDiagnosis))
+            return true;
+        // if year is missing we don't know which census tract is the right one to use, so just check both
+        if (isUnknownYear(_yearOfDiagnosis))
+            return isInvalidCensusTract(getCensusTract2010()) || isInvalidCensusTract(getCensusTract2020());
+        // year isn't missing, so check the census tract we'll end up using for the lookup
+        int dxYear = Integer.parseInt(_yearOfDiagnosis);
+        CensusTract censusTract = dxYear <= YOST_ACS_CENSUS_TRACT_PIVOT_YEAR ? CensusTract.CENSUS_2010 : CensusTract.CENSUS_2020;
+        return isInvalidCensusTract(getCensusTractVariable(censusTract));
+    }
+
+    public boolean hasUnknownStateCountyCensusTractOrYear() {
+        if (isUnknownState(getAddressAtDxState()) || isUnknownCounty(getCountyAtDxAnalysis()) || isUnknownYear(_yearOfDiagnosis))
+            return true;
+        // if year is invalid we don't know which census tract is the right one to use, so just check both
+        if (isInvalidYear(_yearOfDiagnosis))
+            return isUnknownCensusTract(getCensusTract2010()) || isUnknownCensusTract(getCensusTract2020());
+        // year isn't invalid, so check the census tract we'll end up using for the lookup
+        int dxYear = Integer.parseInt(_yearOfDiagnosis);
+        CensusTract censusTract = dxYear <= YOST_ACS_CENSUS_TRACT_PIVOT_YEAR ? CensusTract.CENSUS_2010 : CensusTract.CENSUS_2020;
+        return isUnknownCensusTract(getCensusTractVariable(censusTract));
+    }
+
+    public static boolean isInvalidYear(String year) {
+        return !(isValidYear(year) || isUnknownYear(year));
+    }
+
+    public static boolean isUnknownYear(String year) {
+        return MISSING_YEAR.equals(year);
+    }
+
+    // this is private because it's meant to be more of a helper function
+    // I want users of this class to use the "Invalid" and "Unknown" functions
+    private static boolean isValidYear(String year) {
+        return year != null && year.length() == 4 && NumberUtils.isDigits(year);
+    }
+
 }