+++ src/test/java/com/imsweb/algorithms/yostacspoverty/YostAcsPovertyUtilsTest.java
@@ -11,27 +11,234 @@ public class YostAcsPovertyUtilsTest {
     @Test
     public void testYostAcsPoverty() {
         YostAcsPovertyInputDto idto = new YostAcsPovertyInputDto();
+        YostAcsPovertyOutputDto odto;
+
+        // invalid state
+        idto.setAddressAtDxState("12345");
+        idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2021");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("A", odto.getYostQuintileUS());
+        Assert.assertEquals("A", odto.getYostQuintileState());
+        Assert.assertEquals("A", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("A", odto.getAcsPctPovWhite());
+        Assert.assertEquals("A", odto.getAcsPctPovBlack());
+        Assert.assertEquals("A", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("A", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("A", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("A", odto.getAcsPctPovHispanic());
+
+        // invalid county
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("ABCDE");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2021");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("A", odto.getYostQuintileUS());
+        Assert.assertEquals("A", odto.getYostQuintileState());
+        Assert.assertEquals("A", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("A", odto.getAcsPctPovWhite());
+        Assert.assertEquals("A", odto.getAcsPctPovBlack());
+        Assert.assertEquals("A", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("A", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("A", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("A", odto.getAcsPctPovHispanic());
+
+        // invalid 2010 census tract
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("001");
+        idto.setCensusTract2010("ABCDEF");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2012");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("A", odto.getYostQuintileUS());
+        Assert.assertEquals("A", odto.getYostQuintileState());
+        Assert.assertEquals("A", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("A", odto.getAcsPctPovWhite());
+        Assert.assertEquals("A", odto.getAcsPctPovBlack());
+        Assert.assertEquals("A", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("A", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("A", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("A", odto.getAcsPctPovHispanic());
+
+        // invalid 2020 census tract
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("001");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("ABCDEF");
+        idto.setDateOfDiagnosis("2020");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("A", odto.getYostQuintileUS());
+        Assert.assertEquals("A", odto.getYostQuintileState());
+        Assert.assertEquals("A", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("A", odto.getAcsPctPovWhite());
+        Assert.assertEquals("A", odto.getAcsPctPovBlack());
+        Assert.assertEquals("A", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("A", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("A", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("A", odto.getAcsPctPovHispanic());
+
+        // invalid date of dx
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("001");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("ABCDEF");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("A", odto.getYostQuintileUS());
+        Assert.assertEquals("A", odto.getYostQuintileState());
+        Assert.assertEquals("A", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("A", odto.getAcsPctPovWhite());
+        Assert.assertEquals("A", odto.getAcsPctPovBlack());
+        Assert.assertEquals("A", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("A", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("A", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("A", odto.getAcsPctPovHispanic());
+
+        // county not reported (ie, 000)
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("000");
+        idto.setCensusTract2010("");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2020");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("B", odto.getYostQuintileUS());
+        Assert.assertEquals("B", odto.getYostQuintileState());
+        Assert.assertEquals("B", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("B", odto.getAcsPctPovWhite());
+        Assert.assertEquals("B", odto.getAcsPctPovBlack());
+        Assert.assertEquals("B", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("B", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("B", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("B", odto.getAcsPctPovHispanic());
+
+        // blank state
+        idto.setAddressAtDxState("");
+        idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2018");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("D", odto.getYostQuintileUS());
+        Assert.assertEquals("D", odto.getYostQuintileState());
+        Assert.assertEquals("D", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("D", odto.getAcsPctPovWhite());
+        Assert.assertEquals("D", odto.getAcsPctPovBlack());
+        Assert.assertEquals("D", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("D", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("D", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("D", odto.getAcsPctPovHispanic());
+
+        // blank county
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2018");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("D", odto.getYostQuintileUS());
+        Assert.assertEquals("D", odto.getYostQuintileState());
+        Assert.assertEquals("D", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("D", odto.getAcsPctPovWhite());
+        Assert.assertEquals("D", odto.getAcsPctPovBlack());
+        Assert.assertEquals("D", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("D", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("D", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("D", odto.getAcsPctPovHispanic());
+
+        // blank census tract 2010
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("2012");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("D", odto.getYostQuintileUS());
+        Assert.assertEquals("D", odto.getYostQuintileState());
+        Assert.assertEquals("D", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("D", odto.getAcsPctPovWhite());
+        Assert.assertEquals("D", odto.getAcsPctPovBlack());
+        Assert.assertEquals("D", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("D", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("D", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("D", odto.getAcsPctPovHispanic());
+
+        // blank census tract 2020
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("");
+        idto.setDateOfDiagnosis("2018");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("D", odto.getYostQuintileUS());
+        Assert.assertEquals("D", odto.getYostQuintileState());
+        Assert.assertEquals("D", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("D", odto.getAcsPctPovWhite());
+        Assert.assertEquals("D", odto.getAcsPctPovBlack());
+        Assert.assertEquals("D", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("D", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("D", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("D", odto.getAcsPctPovHispanic());
+
+        // blank date of diagnosis
+        idto.setAddressAtDxState("AK");
+        idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("001702");
+        idto.setDateOfDiagnosis("");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("D", odto.getYostQuintileUS());
+        Assert.assertEquals("D", odto.getYostQuintileState());
+        Assert.assertEquals("D", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("D", odto.getAcsPctPovWhite());
+        Assert.assertEquals("D", odto.getAcsPctPovBlack());
+        Assert.assertEquals("D", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("D", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("D", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("D", odto.getAcsPctPovHispanic());
 
         // 2005 -> no data available
         idto.setAddressAtDxState("AL");
         idto.setCountyAtDxAnalysis("001");
         idto.setCensusTract2010("020200");
+        idto.setCensusTract2020("");
         idto.setDateOfDiagnosis("2005");
-        YostAcsPovertyOutputDto odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
-        Assert.assertNull(odto.getYostQuintileUS());
-        Assert.assertNull(odto.getYostQuintileState());
-        Assert.assertNull(odto.getAcsPctPovAllRaces());
-        Assert.assertNull(odto.getAcsPctPovWhite());
-        Assert.assertNull(odto.getAcsPctPovBlack());
-        Assert.assertNull(odto.getAcsPctPovAIAN());
-        Assert.assertNull(odto.getAcsPctPovAsianNHOPI());
-        Assert.assertNull(odto.getAcsPctPovWhiteNonHisp());
-        Assert.assertNull(odto.getAcsPctPovHispanic());
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("C", odto.getYostQuintileUS());
+        Assert.assertEquals("C", odto.getYostQuintileState());
+        Assert.assertEquals("C", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("C", odto.getAcsPctPovWhite());
+        Assert.assertEquals("C", odto.getAcsPctPovBlack());
+        Assert.assertEquals("C", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("C", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("C", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("C", odto.getAcsPctPovHispanic());
 
-        // 2007 using census2010 with year 2008
+        // test a SEER recode
+        idto.setAddressAtDxState("LO");
+        idto.setCountyAtDxAnalysis("071");
+        idto.setCensusTract2010("007903");
+        idto.setCensusTract2020("");
+        idto.setDateOfDiagnosis("2010");
+        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
+        Assert.assertEquals("5", odto.getYostQuintileUS());
+        Assert.assertEquals("4", odto.getYostQuintileState());
+        Assert.assertEquals("3.03", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("3.81", odto.getAcsPctPovWhite());
+        Assert.assertEquals("4.71", odto.getAcsPctPovBlack());
+        Assert.assertEquals("", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("0.00", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("4.93", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("0.70", odto.getAcsPctPovHispanic());
+
+        // 2007 using census2010 with yearForLookup 2008
         idto.setAddressAtDxState("AL");
         idto.setCountyAtDxAnalysis("001");
         idto.setCensusTract2010("020200");
+        idto.setCensusTract2020("");
         idto.setDateOfDiagnosis("2007");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
         Assert.assertEquals("2", odto.getYostQuintileUS());
@@ -48,6 +255,7 @@ public class YostAcsPovertyUtilsTest {
         idto.setAddressAtDxState("AL");
         idto.setCountyAtDxAnalysis("001");
         idto.setCensusTract2010("020200");
+        idto.setCensusTract2020("");
         idto.setDateOfDiagnosis("2008");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
         Assert.assertEquals("2", odto.getYostQuintileUS());
@@ -64,6 +272,7 @@ public class YostAcsPovertyUtilsTest {
         idto.setAddressAtDxState("AL");
         idto.setCountyAtDxAnalysis("001");
         idto.setCensusTract2010("020200");
+        idto.setCensusTract2020("");
         idto.setDateOfDiagnosis("2010");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
         Assert.assertEquals("2", odto.getYostQuintileUS());
@@ -80,6 +289,7 @@ public class YostAcsPovertyUtilsTest {
         idto.setAddressAtDxState("AK");
         idto.setCountyAtDxAnalysis("020");
         idto.setCensusTract2010("001702");
+        idto.setCensusTract2020("");
         idto.setDateOfDiagnosis("20150101");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
         Assert.assertEquals("5", odto.getYostQuintileUS());
@@ -92,39 +302,10 @@ public class YostAcsPovertyUtilsTest {
         Assert.assertEquals("3.23", odto.getAcsPctPovWhiteNonHisp());
         Assert.assertEquals("0.00", odto.getAcsPctPovHispanic());
 
-        // 2016 using census2020 with year 2018
-        idto.setAddressAtDxState("AK");
-        idto.setCountyAtDxAnalysis("020");
-        idto.setCensusTract2010("001702");
-        idto.setDateOfDiagnosis("2016");
-        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
-        Assert.assertNull(odto.getYostQuintileState());
-        Assert.assertNull(odto.getAcsPctPovAllRaces());
-        Assert.assertNull(odto.getAcsPctPovWhite());
-        Assert.assertNull(odto.getAcsPctPovBlack());
-        Assert.assertNull(odto.getAcsPctPovAIAN());
-        Assert.assertNull(odto.getAcsPctPovAsianNHOPI());
-        Assert.assertNull(odto.getAcsPctPovWhiteNonHisp());
-        Assert.assertNull(odto.getAcsPctPovHispanic());
-
-        // 2018 using census2020
-        idto.setAddressAtDxState("AK");
-        idto.setCountyAtDxAnalysis("020");
-        idto.setCensusTract2010("001702");
-        idto.setDateOfDiagnosis("2018");
-        odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
-        Assert.assertNull(odto.getYostQuintileState());
-        Assert.assertNull(odto.getAcsPctPovAllRaces());
-        Assert.assertNull(odto.getAcsPctPovWhite());
-        Assert.assertNull(odto.getAcsPctPovBlack());
-        Assert.assertNull(odto.getAcsPctPovAIAN());
-        Assert.assertNull(odto.getAcsPctPovAsianNHOPI());
-        Assert.assertNull(odto.getAcsPctPovWhiteNonHisp());
-        Assert.assertNull(odto.getAcsPctPovHispanic());
-
         // 2020 using census2020
         idto.setAddressAtDxState("AK");
         idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("");
         idto.setCensusTract2020("001702");
         idto.setDateOfDiagnosis("2020");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
@@ -141,7 +322,8 @@ public class YostAcsPovertyUtilsTest {
         // 2021 using census2020
         idto.setAddressAtDxState("AK");
         idto.setCountyAtDxAnalysis("020");
-        idto.setCensusTract2010("001702");
+        idto.setCensusTract2010("");
+        idto.setCensusTract2020("001702");
         idto.setDateOfDiagnosis("2021");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
         Assert.assertEquals("4", odto.getYostQuintileUS());
@@ -154,20 +336,22 @@ public class YostAcsPovertyUtilsTest {
         Assert.assertEquals("3.09", odto.getAcsPctPovWhiteNonHisp());
         Assert.assertEquals("8.09", odto.getAcsPctPovHispanic());
 
-        // 2025 using census2020 with year 2021
+        // 2025 using census2020 with yearForLookup 2022
+        // note: this may fail once 2022 data is added to the big SEER data file
         idto.setAddressAtDxState("AK");
         idto.setCountyAtDxAnalysis("020");
+        idto.setCensusTract2010("");
         idto.setCensusTract2020("001702");
         idto.setDateOfDiagnosis("2025");
         odto = YostAcsPovertyUtils.computeYostAcsPovertyData(idto);
-        Assert.assertEquals("4", odto.getYostQuintileUS());
-        Assert.assertEquals("5", odto.getYostQuintileState());
-        Assert.assertEquals("7.12", odto.getAcsPctPovAllRaces());
-        Assert.assertEquals("3.03", odto.getAcsPctPovWhite());
-        Assert.assertEquals("13.46", odto.getAcsPctPovBlack());
-        Assert.assertEquals("34.10", odto.getAcsPctPovAIAN());
-        Assert.assertEquals("0.00", odto.getAcsPctPovAsianNHOPI());
-        Assert.assertEquals("3.09", odto.getAcsPctPovWhiteNonHisp());
-        Assert.assertEquals("8.09", odto.getAcsPctPovHispanic());
+        Assert.assertEquals("C", odto.getYostQuintileUS());
+        Assert.assertEquals("C", odto.getYostQuintileState());
+        Assert.assertEquals("C", odto.getAcsPctPovAllRaces());
+        Assert.assertEquals("C", odto.getAcsPctPovWhite());
+        Assert.assertEquals("C", odto.getAcsPctPovBlack());
+        Assert.assertEquals("C", odto.getAcsPctPovAIAN());
+        Assert.assertEquals("C", odto.getAcsPctPovAsianNHOPI());
+        Assert.assertEquals("C", odto.getAcsPctPovWhiteNonHisp());
+        Assert.assertEquals("C", odto.getAcsPctPovHispanic());
     }
 }